Space decorator in order to bring gaps between messages
