+++ stream-chat-android-ui-components-sample/src/main/res/layout/fragment_deleted_message_component_browser.xml
@@ -9,6 +9,7 @@
         android:id="@+id/recycler_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:layout_margin="8dp"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"/>
 
 </FrameLayout>
