+++ stream-chat-android-ui-components/src/main/res/layout/stream_item_message_deleted.xml
@@ -1,20 +1,42 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content">
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
+    >
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+    <Space
+        android:id="@+id/standard_gap"
+        android:layout_width="match_parent"
+        android:layout_height="2dp"
+        android:visibility="gone"
+        />
+
+    <Space
+        android:id="@+id/big_gap"
+        android:layout_width="match_parent"
+        android:layout_height="8dp"
+        android:visibility="gone"
+        />
+
+    </LinearLayout>
 
     <TextView
         android:id="@+id/deleteLabel"
-        android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_gravity="end"
         android:paddingLeft="@dimen/stream_spacing_medium"
+        android:paddingTop="@dimen/stream_spacing_small"
         android:paddingRight="@dimen/stream_spacing_medium"
         android:paddingBottom="@dimen/stream_spacing_small"
-        android:paddingTop="@dimen/stream_spacing_small"
-        android:layout_gravity="end"
+        android:text="@string/stream_message_deleted_label"
+        android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
         android:textColor="@color/stream_black_50"
-        android:text="@string/stream_message_deleted_label"/>
+        />
 
-</FrameLayout>
+</LinearLayout>
